Create the homepage and first two recipe pages
The homepage was created with 2 lists featuring links.

The first list is recipes of mine, with clickable links for each recipe.
These links take the user to the specified recipe page within my own
website.

The second list is recipes from other chefs, with clickable links for
each recipe. These links take the user to that chef's recipe page.